Customer_Support/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
pydantic==2.6.3
sentence-transformers==5.0.0
chromadb==1.0.15
python-dotenv==1.0.0
scikit-learn==1.6.1
python-multipart==0.0.9
requests==2.32.3 